// src/lib/types/enum/ECircleDomainId.ts
export enum ECircleDomainId {
    Ethereum = 0,
    Avalanche = 1,
    Optimism = 2,
    Arbitrum = 3,
    Solana = 5,
    Base = 6,
    Polygon = 7,
    Unichain = 10,
    Linea = 11,
    Codex = 12,
    Sonic = 13,
    World = 14,
    Monad = 15,
    Sei = 16,
    Binance = 17,
    XDC = 18,
    HyperEVM = 19,
    Ink = 21,
    Plume = 22,
    Starknet = 25,
    Arc = 26,
    Stellar = 27,
    EDGE = 28,
    Injective = 29,
    Morph = 30,
    Pharos = 31,
    Cronos = 32,
    XLayer = 37,
}